Более логичное разбиение на файлы
